+++ CHANGELOG.txt
@@ -1,5 +1,8 @@
 2.20.5:
   - Fix cdrom_image type instead of default in plugin yaml.
+  - add support of changing boot order during vm creation
+  - boot_order, disk_keys, ethernet_keys to properties of node
+    cloudify.nodes.vsphere.OvfDeployment & cloudify.nodes.vsphere.Server
 2.20.4:
   - Add change_boot_order operation and new ISO type -upload ISO to datastore-.
 2.20.3:
